Make coverImage() return more information
This changes the coverImage() method so that it returns the full
manifest item as opposed to just the name of the file. This is
important because some applications will need to access the associated
content type properties so that they can tell whether or not the
image type is supported.